// src/app/core/models/post.interface.ts
export interface Post {

    _id: string;

    body: string;

    image?: string;

    privacy: string;

    user: User;

    sharedPost: null;

    likes: any[];

    createdAt: Date;

    commentsCount: number;

    topComment: null;

    sharesCount: number;

    likesCount: number;

    isShare: boolean;

    id: string;

    bookmarked: boolean;

}
export interface User {

    _id: string;
    name: string;
    username: string;
    photo: string;
}
export interface LikeResponse {
    data: {
        post: {
            likes: string[];
        };
    };
}